Fixing problems after last merge
